# `Development`: Improve notifications performance