--- a/src/main/java/de/tum/in/www1/artemis/repository/GroupNotificationRepository.java
+++ b/src/main/java/de/tum/in/www1/artemis/repository/GroupNotificationRepository.java
@@ -6,13 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import de.tum.in.www1.artemis.domain.notification.GroupNotification;
-import de.tum.in.www1.artemis.domain.notification.Notification;
 
 /**
  * Spring Data repository for the Notification entity.
  */
 @Repository
-public interface GroupNotificationRepository extends JpaRepository<Notification, Long> {
+public interface GroupNotificationRepository extends JpaRepository<GroupNotification, Long> {
 
     List<GroupNotification> findAllByCourseId(Long courseId);
 }