--- a/src/main/java/de/tum/in/www1/artemis/repository/SystemNotificationRepository.java
+++ b/src/main/java/de/tum/in/www1/artemis/repository/SystemNotificationRepository.java
@@ -16,6 +16,12 @@ import de.tum.in.www1.artemis.domain.notification.SystemNotification;
 @Repository
 public interface SystemNotificationRepository extends JpaRepository<SystemNotification, Long> {
 
-    @Query("SELECT distinct notification FROM SystemNotification notification WHERE (notification.expireDate >= :now OR notification.expireDate IS NULL) ORDER BY notification.notificationDate ASC")
+    @Query("""
+            SELECT DISTINCT notification
+            FROM SystemNotification notification
+            WHERE notification.expireDate >= :now
+                OR notification.expireDate IS NULL
+            ORDER BY notification.notificationDate ASC
+            """)
     List<SystemNotification> findAllActiveAndFutureSystemNotifications(@Param("now") ZonedDateTime now);
 }