--- a/src/main/resources/config/liquibase/changelog/20230812135300_changelog.xml
+++ b/src/main/resources/config/liquibase/changelog/20230812135300_changelog.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+                   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                   xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-latest.xsd">
+    <changeSet author="krusche" id="20230812135300">
+        <createIndex tableName="notification" indexName="notification_discriminator">
+            <column name="discriminator"/>
+        </createIndex>
+        <createIndex tableName="notification" indexName="notification_expire_date">
+            <column name="expire_date"/>
+        </createIndex>
+    </changeSet>
+</databaseChangeLog>